R User Group Introduction Slides

This function creates an ioslides R Markdown template, with default content routinely used when introducing Chicago R user group meetups.

intro_slides(...)

Arguments

...
Arguments passed to rmarkdown::ioslides_presentation function.

Source

See R Studio's Document Templates for more information on creating custom R Markdown templates for your meetup.

Value

A folder containing the R Markdown template as well as supporting images.

Details

Slides are useful because they look good and you won't forget to do important things like thanking the Sponsors! Slides can be accessed from within R Studio using the New R Markdown dialog menu, and then selecting From Template. One can also use the draft function, exemplified in the Examples section below.

Default template includes placeholder slides for the following:

Local R news: Announcements about your local meetup or general R community.

Global R news: R Consortium announcements and upcoming conferences like R/Finance and UseR!.

R-Ladies: Welcome the local R-Ladies chapter.

Jobs: Ask the audience if anyone is hiring or if recuiters are present.

Thank the Sponsors: Display their logo and promotional material of preference.

R Swag: Giveaways of hex stickers, Raffle off books, etc.

Tonights presentations: Summary of presentations.

Examples

# NOT RUN {
rmarkdown::draft(file="Meetup-2018-n5.Rmd", template = "intro_slides", package="RUGtools")
# }