Estimate your pizza order

Function to compute how many pizzas to order based on arguments provided. Defaults to thin crust.

pizza_estimate(registered = 120, pizza_diameter = 18, attend_rate = 0.6,
  serving = 2, style = "thin")

Arguments

registered
Integer. The number of people registered on Meetup.com

pizza_diameter
Integer. Diameter of the largest pizza available in inches. Defaults to 18 inches for an extra large pizza.

attend_rate
Numeric of type double. The percentage estimate of attendance. Defaults to 60 percent attendance rate, expressed as 0.60.

serving
Integer or Numeric. Number of servings per person. Default is 2 and does not have to be a whole number. See references for methodology!

style
Character vector of either "thin", "pan", or "deepdish". Defaults to "thin" crust, please read the Details section for important information. Most pizza parlors have "thin" and "pan" styles..."deepdish" is an option because this function was written in Chicago, though it is our sincerest hope, that we shall never set this parameter.

Value

pizza_estimate returns a data.frame containing 5 variables. These are registered, est_attend, eaters_per_pizza, style, and the pizza_estimate.

Details

Chicago thin crust is cut into small squares, known as "tavern style" or "party cut".

Channeling our ever-curious pizza scientist, it turns out, the "party cut" inherits a few very attractive properties when dividing multiple pizzas among n guests. Smaller pieces allow guests to more precisely estimate pizza consumption, thus decreasing the integer-programming problem exacerbated by larger slices.

By requesting the "party cut", one reduces the tragedy that is wasted pizza and exhibits great stewardship of the sponsor resources entrusted to them.

References

Lisa R. Young, Ph.D., R.D. Nestle USA Site: Enjoy Pizza and PDF: The Fun of Pizza, the Balance of Good Nutrition, page 5.

Examples

pizza_estimate(registered=150, pizza_diameter=18, attend_rate=0.60, serving = 2,
style = "thin")

#>   registered est_attend eaters_per_pizza style pizza_estimate
#> 1        150         90         5.342811  thin             17